Improve regex performance.
Previous approach was to:
  -> Run multiple regexes on every clang result
  -> Filter down matches that do not match current line.

New approach is:
  -> Filter down matches that do not match current line
  -> Run multiple regexes on these matches.

In my tests, worst case senarios (single letter completes on
global scope), went from ~100ms to ~10ms.  This is only a
*small* performance improvement because most of the time
spent in auto-completing a line is in the clang executable
(for my simple cases clang was taking ~250ms to ~850ms per
request).